View Mapping 추가
