+++ PlayGround/.settings/org.eclipse.wst.common.project.facet.core.xml
@@ -2,6 +2,6 @@
 <faceted-project>
   <fixed facet="jst.web"/>
   <fixed facet="jst.java"/>
-  <installed facet="jst.web" version="2.5"/>
   <installed facet="jst.java" version="1.8"/>
+  <installed facet="jst.web" version="3.1"/>
 </faceted-project>
