+++ PlayGround/src/main/java/com/controller/news/NewsController.java
@@ -1,23 +1,14 @@
 package com.controller.news;
 
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
 
-import com.dto.NewsDTO;
-import com.service.NewsService;
-
-@RestController
-@RequestMapping("/news/")
+@Controller
+@RequestMapping("/news")
 public class NewsController {
-	@Autowired
-	NewsService nService;
-	
-	@GetMapping("/newsList")
-	public List<NewsDTO> newsMainList() {
-		return nService.newsSelect();
+	@GetMapping("/list")
+	public String newsList() {
+		return "News/newsList";
 	}
 }
