+++ PlayGround/src/main/webapp/WEB-INF/views/common/header.jsp
@@ -23,7 +23,7 @@
 			<nav>
 				<div class="empty"></div>
 				<ul class="nav">
-					<li><a href="${pageContext.request.contextPath}/Board/boardList.jsp">게시판</a></li>
+					<li><a href="board/boardList">게시판</a></li>
 					<li><a href="${pageContext.request.contextPath}/Board/newsList.jsp">뉴스 및 소식</a></li>
 					<li><a href="${pageContext.request.contextPath}/MyPageServlet">마이페이지</a></li>
 					<li><a href="${pageContext.request.contextPath}/LogoutServlet">로그아웃</a></li>
@@ -36,7 +36,7 @@
 			<nav>
 				<div class="empty"></div>
 				<ul class="nav">
-					<li><a href="${pageContext.request.contextPath}/Board/boardList.jsp">게시판</a></li>
+					<li><a href="board/boardList">게시판</a></li>
 					<li><a href="${pageContext.request.contextPath}/Board/newsList.jsp">뉴스 및 소식</a></li>
 					<li><a href="LoginForm">로그인</a></li>
 					<li><a href="${pageContext.request.contextPath}/MemberUIServlet">회원가입</a></li>
