+++ Servers/.settings/org.eclipse.wst.server.core.prefs
@@ -1,3 +1,3 @@
-#Thu Jun 03 15:55:28 KST 2021
+#Fri Jun 04 12:01:28 KST 2021
 org.eclipse.wst.server.core.isServerProject=true
 eclipse.preferences.version=1
